Feat: Entity 폴더 이동 및 수정
가격, 포인트 타입 int에서 BigDecimal로 변경, 카테고리 Entity 카테고리 이름 외 나머지 정보 옵션 Entity로 이동을 위해 제거

+++ src/main/java/gift/dto/pointDto/PointDto.java
@@ -1,6 +1,6 @@
 package gift.dto.pointDto;
 
-import gift.model.member.Member;
+import java.math.BigDecimal;
 
-public record PointDto(Long MemberId, Integer points) {
+public record PointDto(Long MemberId, BigDecimal points) {
 }
