+++ src/main/java/gift/dto/productDto/ProductResponseDto.java
@@ -1,8 +1,10 @@
 package gift.dto.productDto;
 
+import java.math.BigDecimal;
+
 public record ProductResponseDto(Long id,
                                  String name,
-                                 int price,
+                                 BigDecimal price,
                                  String imageUrl,
                                  Long categoryId) { }
